refactor: rename Android component file to avoid a bug in Codegen
